requirements.txt
# MathRAG-Gate Dependencies
# Python 3.10+ recommended

# Core Framework
llama-index==0.10.0
llama-index-core==0.10.0
llama-index-llms-ollama==0.1.0
llama-index-embeddings-huggingface==0.2.0
llama-index-retrievers-bm25==0.2.0

# Deep Learning & Embeddings
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Data Processing
datasets>=2.14.0
huggingface-hub>=0.19.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & Environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Progress & Logging
tqdm>=4.65.0

# Optional: For better performance
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA available

# Development & Testing (Optional)
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
